+++ src/components/ToolbarDivider.tsx
@@ -1,4 +1,4 @@
-import Divider from '@mui/material/Divider';
+import Divider, { type DividerProps } from '@mui/material/Divider';
 
 import { makeStyles } from '@skybrush/app-theme-mui';
 
@@ -10,7 +10,9 @@ const useStyles = makeStyles((theme) => ({
   },
 }));
 
-export const ToolbarDivider = (props) => {
+type Props = Omit<DividerProps, 'className'>;
+
+export const ToolbarDivider = (props: Props) => {
   const classes = useStyles();
   return <Divider className={classes.divider} {...props} />;
 };