+++ src/features/settings/selectors.d.ts
@@ -1,11 +1,12 @@
 import type { APIKeysRecord } from '~/APIKeys';
+import type { UAVFilter } from '~/model/filtering';
+import type { UAVOperationConfirmationStyle } from '~/model/settings';
 import type { RootState } from '~/store/reducers';
 import type {
   UAVListLayout,
   UAVListOrientation,
   UAVSortKeyAndOrder,
 } from './types';
-import type { UAVFilter } from '~/model/filtering';
 
 export function shouldOptimizeUIForTouch(state: any): boolean;
 export function getAPIKeys(state: RootState): APIKeysRecord;
@@ -20,3 +21,6 @@ export function isShowingEmptyMissionSlots(state: RootState): boolean;
 export function isShowingMissionIds(state: RootState): boolean;
 export function getMinimumIndoorTakeoffSpacing(state: RootState): number;
 export function getMinimumOutdoorTakeoffSpacing(state: RootState): number;
+export function getUAVOperationConfirmationStyle(
+  state: RootState
+): UAVOperationConfirmationStyle;