+++ src/features/show/slice.ts
@@ -13,16 +13,17 @@ import isNil from 'lodash-es/isNil';
 import set from 'lodash-es/set';
 
 import { type Clock } from '~/features/clocks/types';
+import type { Schedule } from '~/flockwave/schedule';
 import type UAV from '~/model/uav';
 import { type LonLat } from '~/utils/geography';
 import { type Coordinate3D } from '~/utils/math';
 import { noPayload } from '~/utils/redux';
 
 import {
-  type AltitudeReferenceSpecification,
   DEFAULT_ALTITUDE_REFERENCE,
   DEFAULT_ROOM_SIZE,
   DEFAULT_TAKEOFF_HEADING,
+  type AltitudeReferenceSpecification,
   type TakeoffHeadingSpecification,
 } from './constants';
 import {
@@ -98,6 +99,12 @@ type ShowSliceState = {
     /** Whether the state variables in this object are synced with the server */
     syncStatusWithServer: SettingsSynchronizationStatus;
   };
+
+  /**
+   * The schedule of the collective RTH being executed or `undefined` if
+   * collective RTH was not triggered.
+   */
+  collectiveRTHSchedule?: Schedule;
 
   startTimeDialog: {
     open: boolean;
@@ -188,6 +195,8 @@ const initialState: ShowSliceState = {
     syncStatusWithServer: SettingsSynchronizationStatus.NOT_SYNCED,
   },
 
+  collectiveRTHSchedule: undefined,
+
   startTimeDialog: {
     open: false,
   },
@@ -218,6 +227,9 @@ const { actions, reducer } = createSlice({
 
       // Last upload result cleared in the upload feature as it also handles
       // this action
+
+      // Reset collective RTH schedule
+      state.collectiveRTHSchedule = undefined;
     }),
 
     clearManualPreflightChecks: noPayload<ShowSliceState>((state) => {
@@ -308,6 +320,9 @@ const { actions, reducer } = createSlice({
       // Just in case the "show changed" warning was triggered while we tried
       // to load it...
       state.changedSinceLoaded = false;
+
+      // Reset collective RTH schedule
+      state.collectiveRTHSchedule = undefined;
     },
 
     loadingPromiseRejected(state) {
@@ -360,6 +375,10 @@ const { actions, reducer } = createSlice({
     revokeTakeoffAreaApproval: noPayload<ShowSliceState>((state) => {
       state.preflight.takeoffAreaApprovedAt = undefined;
     }),
+
+    setCollectiveRTHSchedule(state, action: PayloadAction<Schedule>) {
+      state.collectiveRTHSchedule = action.payload;
+    },
 
     setEnvironmentType(state, action: PayloadAction<EnvironmentType>) {
       state.environment.type = action.payload;
@@ -504,6 +523,9 @@ const { actions, reducer } = createSlice({
           state.start.utcTime = undefined;
         }
       }
+
+      // Reset collective RTH schedule
+      state.collectiveRTHSchedule = undefined;
     },
 
     setUAVIdsToStartAutomatically(
@@ -557,6 +579,7 @@ export const {
   openStartTimeDialog,
   openTakeoffAreaSetupDialog,
   revokeTakeoffAreaApproval,
+  setCollectiveRTHSchedule,
   setEnvironmentType,
   setIndoorShowOrientation,
   setIndoorShowTakeoffHeadingSpecification,